chore(space-wizard): rename locals and file-name
