+++ src/app/space-wizard/services/app-generator.service.ts
@@ -16,7 +16,7 @@ export {
 } from './contracts/app-generator-service';
 
 export { IAppGeneratorServiceProvider, FieldSetServiceProvider } from './providers/app-generator-service.provider';
-export { AppGeneratorConfigurationService } from './concrete/field-lookup.service';
+export { AppGeneratorConfigurationService } from './concrete/app-generator-configuration.service';
