+++ src/app/space-wizard/services/concrete/app-generator-configuration.service.ts
